Updated register definition for WDOG_WCR_WT
It seems the header definition for WDOG_WCR_WT has changed. The reader imxrt.h that ships with arduinoteensy 1.51 defines the register like this:

#define WDOG_WCR_WT(n)				((uint16_t)(((n) & 0xFF) << 8))

I could compile the code again and have verified, that the watch dog triggers when I neglegt feeding it for 10s (given it has been initialized with setTimeout(10.0f)).